ultimo commit
Problemas com o projeto:
1- Não consegui implementar a sintaxe de PK e FK;
2- Documentação com swagger não funcional

+++ LabShoppingList/src/main/java/futuroDev/Projeto03/LabShoppingList/models/Produto.java
@@ -10,11 +10,13 @@ public class Produto {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    @ManyToOne
+//    @JoinColumn(name = "produto", referencedColumnName = "id")
     private long id;
     private String nome;
     private String descricao;
 
-    private int valor;
+    private double valor;
 
     private String status;
 
@@ -42,11 +44,11 @@ public class Produto {
         this.descricao = descricao;
     }
 
-    public int getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
